+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/base/activity/BaseActivity.kt
@@ -16,7 +16,7 @@ import es.dmoral.toasty.Toasty
 import io.reactivex.disposables.CompositeDisposable
 import mustafaozhan.github.com.mycurrencies.R
 import mustafaozhan.github.com.mycurrencies.base.viewmodel.BaseViewModel
-import mustafaozhan.github.com.mycurrencies.extension.getImageResourceByName
+import mustafaozhan.github.com.mycurrencies.function.extension.getImageResourceByName
 import java.util.Locale
 import javax.inject.Inject
