+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/base/viewmodel/BaseViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.disposables.CompositeDisposable
-import mustafaozhan.github.com.mycurrencies.extension.applySchedulers
+import mustafaozhan.github.com.mycurrencies.function.extension.applySchedulers
 
 /**
  * Created by Mustafa Ozhan on 7/10/18 at 9:40 PM on Arch Linux wit Love <3.
