+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/extension/DBMigrationExtension.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.extension
+package mustafaozhan.github.com.mycurrencies.function.extension
 
 import androidx.sqlite.db.SupportSQLiteDatabase
