+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/lambda/IfAllSafeLambda.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.function
+package mustafaozhan.github.com.mycurrencies.function.lambda
 
 inline fun <reified T> ifAllSafe(
     vararg elements: T?,
