+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/scope/MapScpoe.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.function
+package mustafaozhan.github.com.mycurrencies.function.scope
 
 inline fun <reified T, reified R> T.mapTo(
     transform: T.(map: T) -> R?
