+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/scope/SafeScope.kt
@@ -1,0 +1,1 @@
+package mustafaozhan.github.com.mycurrencies.function.scope
