+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/function/scope/WhetherScope.kt
@@ -1,4 +1,4 @@
-package mustafaozhan.github.com.mycurrencies.function
+package mustafaozhan.github.com.mycurrencies.function.scope
 
 inline fun <reified T> T.whether(
     method: T.(condition: T) -> Boolean
