+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/ui/main/fragment/calculator/CalculatorFragment.kt
@@ -11,15 +11,15 @@ import kotlinx.android.synthetic.main.item_currency.view.txt_amount
 import mustafaozhan.github.com.mycurrencies.R
 import mustafaozhan.github.com.mycurrencies.base.fragment.BaseViewBindingFragment
 import mustafaozhan.github.com.mycurrencies.databinding.FragmentCalculatorBinding
-import mustafaozhan.github.com.mycurrencies.extension.addText
-import mustafaozhan.github.com.mycurrencies.extension.checkAd
-import mustafaozhan.github.com.mycurrencies.extension.dropDecimal
-import mustafaozhan.github.com.mycurrencies.extension.reObserve
-import mustafaozhan.github.com.mycurrencies.extension.replaceNonStandardDigits
-import mustafaozhan.github.com.mycurrencies.extension.setBackgroundByName
-import mustafaozhan.github.com.mycurrencies.extension.tryToSelect
-import mustafaozhan.github.com.mycurrencies.function.whether
-import mustafaozhan.github.com.mycurrencies.function.whetherNot
+import mustafaozhan.github.com.mycurrencies.function.extension.addText
+import mustafaozhan.github.com.mycurrencies.function.extension.checkAd
+import mustafaozhan.github.com.mycurrencies.function.extension.dropDecimal
+import mustafaozhan.github.com.mycurrencies.function.extension.reObserve
+import mustafaozhan.github.com.mycurrencies.function.extension.replaceNonStandardDigits
+import mustafaozhan.github.com.mycurrencies.function.extension.setBackgroundByName
+import mustafaozhan.github.com.mycurrencies.function.extension.tryToSelect
+import mustafaozhan.github.com.mycurrencies.function.scope.whether
+import mustafaozhan.github.com.mycurrencies.function.scope.whetherNot
 import mustafaozhan.github.com.mycurrencies.model.Rates
 import mustafaozhan.github.com.mycurrencies.room.AppDatabase
 import mustafaozhan.github.com.mycurrencies.ui.main.fragment.settings.SettingsFragment
