+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/ui/main/fragment/calculator/CalculatorFragmentAdapter.kt
@@ -9,9 +9,9 @@ import kotlinx.android.synthetic.main.item_currency.view.txt_type
 import mustafaozhan.github.com.mycurrencies.R
 import mustafaozhan.github.com.mycurrencies.base.adapter.BaseRecyclerViewAdapter
 import mustafaozhan.github.com.mycurrencies.base.adapter.BaseViewHolder
-import mustafaozhan.github.com.mycurrencies.extension.getFormatted
-import mustafaozhan.github.com.mycurrencies.extension.replaceNonStandardDigits
-import mustafaozhan.github.com.mycurrencies.extension.setBackgroundByName
+import mustafaozhan.github.com.mycurrencies.function.extension.getFormatted
+import mustafaozhan.github.com.mycurrencies.function.extension.replaceNonStandardDigits
+import mustafaozhan.github.com.mycurrencies.function.extension.setBackgroundByName
 import mustafaozhan.github.com.mycurrencies.model.Currency
 
 /**
