+++ app/src/main/java/mustafaozhan/github/com/mycurrencies/ui/main/fragment/calculator/CalculatorViewModel.kt
@@ -9,18 +9,18 @@ import kotlinx.coroutines.launch
 import mustafaozhan.github.com.mycurrencies.base.viewmodel.BaseDataViewModel
 import mustafaozhan.github.com.mycurrencies.data.repository.BackendRepository
 import mustafaozhan.github.com.mycurrencies.data.repository.PreferencesRepository
-import mustafaozhan.github.com.mycurrencies.extension.calculateResult
-import mustafaozhan.github.com.mycurrencies.extension.getFormatted
-import mustafaozhan.github.com.mycurrencies.extension.getThroughReflection
-import mustafaozhan.github.com.mycurrencies.extension.insertInitialCurrencies
-import mustafaozhan.github.com.mycurrencies.extension.removeUnUsedCurrencies
-import mustafaozhan.github.com.mycurrencies.extension.replaceNonStandardDigits
-import mustafaozhan.github.com.mycurrencies.extension.replaceUnsupportedCharacters
-import mustafaozhan.github.com.mycurrencies.extension.toPercent
-import mustafaozhan.github.com.mycurrencies.function.either
-import mustafaozhan.github.com.mycurrencies.function.mapTo
-import mustafaozhan.github.com.mycurrencies.function.whether
-import mustafaozhan.github.com.mycurrencies.function.whetherNot
+import mustafaozhan.github.com.mycurrencies.function.extension.calculateResult
+import mustafaozhan.github.com.mycurrencies.function.extension.getFormatted
+import mustafaozhan.github.com.mycurrencies.function.extension.getThroughReflection
+import mustafaozhan.github.com.mycurrencies.function.extension.insertInitialCurrencies
+import mustafaozhan.github.com.mycurrencies.function.extension.removeUnUsedCurrencies
+import mustafaozhan.github.com.mycurrencies.function.extension.replaceNonStandardDigits
+import mustafaozhan.github.com.mycurrencies.function.extension.replaceUnsupportedCharacters
+import mustafaozhan.github.com.mycurrencies.function.extension.toPercent
+import mustafaozhan.github.com.mycurrencies.function.scope.either
+import mustafaozhan.github.com.mycurrencies.function.scope.mapTo
+import mustafaozhan.github.com.mycurrencies.function.scope.whether
+import mustafaozhan.github.com.mycurrencies.function.scope.whetherNot
 import mustafaozhan.github.com.mycurrencies.model.Currencies
 import mustafaozhan.github.com.mycurrencies.model.Currency
 import mustafaozhan.github.com.mycurrencies.model.CurrencyResponse
