--- app/src/test/java/mustafaozhan/github/com/mycurrencies/constant/Constant.kt
@@ -1,7 +1,0 @@
-package mustafaozhan.github.com.mycurrencies.constant
-
-const val UN_EXPECTED = "Unexpected"
-const val EXPECTED = "Expected"
-
-@Suppress("MaybeConst")
-val SOME_STRING: String? = "Some String"
