--- app/src/test/java/mustafaozhan/github/com/mycurrencies/function/MapScopeTest.kt
@@ -1,49 +1,0 @@
-package mustafaozhan.github.com.mycurrencies.function
-
-import mustafaozhan.github.com.mycurrencies.constant.EXPECTED
-import mustafaozhan.github.com.mycurrencies.constant.SOME_STRING
-import mustafaozhan.github.com.mycurrencies.constant.UN_EXPECTED
-import org.junit.Assert
-import org.junit.Assert.assertEquals
-import org.junit.Assert.assertTrue
-import org.junit.Test
-
-class MapScopeTest : MainScopeTest() {
-    @Test
-    fun mapTo() {
-        subjectScope
-            ?.mapTo { it.trueCondition }
-            ?.whether { it }
-            ?.let { assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
-
-        subjectScope
-            ?.mapTo { it.falseCondition }
-            ?.whether { it }
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { assertTrue(EXPECTED, true) }
-        subjectScope
-            ?.mapTo { trueCondition }
-            ?.whether { it }
-            ?.let { assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
-
-        subjectScope
-            ?.mapTo { falseCondition }
-            ?.whether { it }
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { assertTrue(EXPECTED, true) }
-    }
-
-    @Test
-    fun `extraordinary mapTo`() = subjectScope
-        .mapTo { SOME_STRING }
-        ?.mapTo { it -> it.length }
-        ?.let { assertEquals(11, it) }
-        ?: run { Assert.fail(UN_EXPECTED) }
-            .mapTo { subjectScope?.trueCondition }
-            ?.whether { it }
-            ?.mapTo { !it }
-            ?.let { Assert.fail(UN_EXPECTED) }
-        ?: run { assertTrue(EXPECTED, true) }
-}
