+++ app/src/test/java/mustafaozhan/github/com/mycurrencies/function/model/FunctionTestSubject.kt
@@ -1,6 +1,6 @@
 package mustafaozhan.github.com.mycurrencies.function.model
 
-class ScopeTestSubject {
+class FunctionTestSubject {
     var trueCondition = true
     var falseCondition = false
     var nullAbleCondition: Boolean? = null
