+++ app/src/test/java/mustafaozhan/github/com/mycurrencies/function/scope/EitherScopeTest.kt
@@ -1,7 +1,5 @@
-package mustafaozhan.github.com.mycurrencies.function
+package mustafaozhan.github.com.mycurrencies.function.scope
 
-import mustafaozhan.github.com.mycurrencies.constant.EXPECTED
-import mustafaozhan.github.com.mycurrencies.constant.UN_EXPECTED
 import org.junit.Assert
 import org.junit.Test
 
@@ -9,100 +7,100 @@ class EitherScopeTest : MainScopeTest() {
 
     @Test
     fun `either mix mix`() {
-        subjectScope
+        subjectFunction
             ?.either({ it.trueCondition }, { it.falseCondition })
             ?.either({ trueCondition }, { falseCondition })
-            ?.let { Assert.assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
+            ?.let { Assert.assertTrue(Companion.EXPECTED, true) }
+            ?: run { Assert.fail(Companion.UN_EXPECTED) }
 
-        subjectScope
+        subjectFunction
             ?.either({ it.falseCondition }, { it.trueCondition })
             ?.either({ falseCondition }, { trueCondition })
-            ?.let { Assert.assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
+            ?.let { Assert.assertTrue(Companion.EXPECTED, true) }
+            ?: run { Assert.fail(Companion.UN_EXPECTED) }
     }
 
     @Test
     fun `either mix true`() {
-        subjectScope
+        subjectFunction
             ?.either({ it.trueCondition }, { it.falseCondition })
             ?.either({ trueCondition }, { trueCondition })
-            ?.let { Assert.assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
+            ?.let { Assert.assertTrue(Companion.EXPECTED, true) }
+            ?: run { Assert.fail(Companion.UN_EXPECTED) }
 
-        subjectScope
+        subjectFunction
             ?.either({ it.falseCondition }, { it.trueCondition })
             ?.either({ trueCondition }, { trueCondition })
-            ?.let { Assert.assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
+            ?.let { Assert.assertTrue(Companion.EXPECTED, true) }
+            ?: run { Assert.fail(Companion.UN_EXPECTED) }
     }
 
     @Test
     fun `either mix false`() {
-        subjectScope
+        subjectFunction
             ?.either({ it.trueCondition }, { it.falseCondition })
             ?.either({ falseCondition }, { falseCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
 
-        subjectScope
+        subjectFunction
             ?.either({ it.falseCondition }, { it.trueCondition })
             ?.either({ falseCondition }, { falseCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
     }
 
     @Test
     fun `eitherNot mix`() {
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.trueCondition }, { it.falseCondition })
             ?.eitherNot({ trueCondition }, { falseCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
 
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.falseCondition }, { it.trueCondition })
             ?.eitherNot({ falseCondition }, { trueCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
     }
 
     @Test
     fun `eitherNot mix true`() {
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.trueCondition }, { it.falseCondition })
             ?.eitherNot({ trueCondition }, { trueCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
 
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.falseCondition }, { it.trueCondition })
             ?.eitherNot({ trueCondition }, { trueCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
     }
 
     @Test
     fun `eitherNot mix false`() {
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.trueCondition }, { it.falseCondition })
             ?.eitherNot({ falseCondition }, { falseCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
 
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.falseCondition }, { it.trueCondition })
             ?.eitherNot({ falseCondition }, { falseCondition })
-            ?.let { Assert.fail(UN_EXPECTED) }
-            ?: run { Assert.assertTrue(EXPECTED, true) }
+            ?.let { Assert.fail(Companion.UN_EXPECTED) }
+            ?: run { Assert.assertTrue(Companion.EXPECTED, true) }
     }
 
     @Test
     fun `eitherNot false false`() {
-        subjectScope
+        subjectFunction
             ?.eitherNot({ it.falseCondition }, { it.falseCondition })
             ?.eitherNot({ falseCondition }, { falseCondition })
-            ?.let { Assert.assertTrue(EXPECTED, true) }
-            ?: run { Assert.fail(UN_EXPECTED) }
+            ?.let { Assert.assertTrue(Companion.EXPECTED, true) }
+            ?: run { Assert.fail(Companion.UN_EXPECTED) }
     }
 }
